+++ pkg/build/other/msf/msf.go
@@ -94,36 +94,41 @@ func Build(opts ...Option) error {
 		}
 
 		for _, cve := range cves {
-			y := strings.Split(cve, "-")[1]
-			if err := os.MkdirAll(filepath.Join(options.destDir, y), os.ModePerm); err != nil {
-				return errors.Wrapf(err, "mkdir %s", filepath.Join(options.destDir, y))
-			}
+			if err := func() error {
+				y := strings.Split(cve, "-")[1]
+				if err := os.MkdirAll(filepath.Join(options.destDir, y), os.ModePerm); err != nil {
+					return errors.Wrapf(err, "mkdir %s", filepath.Join(options.destDir, y))
+				}
 
-			df, err := os.OpenFile(filepath.Join(options.destDir, y, fmt.Sprintf("%s.json", cve)), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0644)
-			if err != nil {
-				return errors.Wrapf(err, "open %s", filepath.Join(options.destDir, y, fmt.Sprintf("%s.json", cve)))
-			}
-			defer df.Close()
+				df, err := os.OpenFile(filepath.Join(options.destDir, y, fmt.Sprintf("%s.json", cve)), os.O_RDWR|os.O_CREATE|os.O_APPEND, 0644)
+				if err != nil {
+					return errors.Wrapf(err, "open %s", filepath.Join(options.destDir, y, fmt.Sprintf("%s.json", cve)))
+				}
+				defer df.Close()
 
-			var dv build.Vulnerability
-			if err := json.NewDecoder(df).Decode(&dv); err != nil && !errors.Is(err, io.EOF) {
-				return errors.Wrap(err, "decode json")
-			}
+				var dv build.Vulnerability
+				if err := json.NewDecoder(df).Decode(&dv); err != nil && !errors.Is(err, io.EOF) {
+					return errors.Wrap(err, "decode json")
+				}
 
-			fillVulnerability(&dv, &m, cve)
+				fillVulnerability(&dv, &m, cve)
 
-			if err := df.Truncate(0); err != nil {
-				return errors.Wrap(err, "truncate file")
-			}
-			if _, err := df.Seek(0, 0); err != nil {
-				return errors.Wrap(err, "set offset")
-			}
-			enc := json.NewEncoder(df)
-			enc.SetIndent("", "  ")
-			if err := enc.Encode(dv); err != nil {
-				return errors.Wrap(err, "encode json")
-			}
+				if err := df.Truncate(0); err != nil {
+					return errors.Wrap(err, "truncate file")
+				}
+				if _, err := df.Seek(0, 0); err != nil {
+					return errors.Wrap(err, "set offset")
+				}
+				enc := json.NewEncoder(df)
+				enc.SetIndent("", "  ")
+				if err := enc.Encode(dv); err != nil {
+					return errors.Wrap(err, "encode json")
+				}
 
+				return nil
+			}(); err != nil {
+				return err
+			}
 		}
 		return nil
 	}); err != nil {
